Add unit testing support
The project is reorganized to include a package for the Secret Hitler app, and a test package for that app. The contents of main.py are now in secret_hitler/app.py and main.py is a simple script that will import app and run it.
