+++ secret_hitler/img/LiberalBoard.png
