+++ secret_hitler/img/chancellor.png
