+++ secret_hitler/img/liberal_policy.png
