+++ secret_hitler/img/liberal_role.png
