+++ secret_hitler/img/policy_fascist.png
