+++ secret_hitler/img/policy_liberal.png
